requirements.txt
<<<<<<< HEAD
=======
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
>>>>>>> feat/rag
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
<<<<<<< HEAD
cffi==1.17.1
channels==4.2.0
constantly==23.10.4
cryptography==44.0.2
Django==5.1.7
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
service-identity==24.2.0
setuptools==77.0.3
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==7.2
=======
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
dataclasses-json==0.6.7
Django==5.1.7
djangorestframework==3.15.2
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-ollama==0.3.0
langchain-text-splitters==0.3.7
langsmith==0.3.18
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==1.31.0
numpy==2.2.4
ollama==0.4.7
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pyOpenSSL==25.0.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.23.1
service-identity==24.2.0
setuptools==77.0.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlparse==0.5.3
streamlit==1.43.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
Twisted==24.11.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
zope.interface==7.2
zstandard==0.23.0
>>>>>>> feat/rag
